requirements.txt
Django
git+https://github.com/TheEnbyperor/django-keycloak-auth@9e634ea6866faddc6c67f7a6c89c3f3d28d573c6
sentry-sdk
psycopg2-binary
gunicorn
django-crispy-forms
django-countries
django-phonenumber-field[phonenumbers]
django-xff
grpcio
protobuf
django-grpc
